Implement internal mergeCollectionWithPatches function
