+++ API.md
@@ -28,8 +28,8 @@ Values of type <code>Object</code> get merged with the old value, whilst for <co
 applied in the order they were called. Note: <code>Onyx.set()</code> calls do not work this way so use caution when mixing
 <code>Onyx.merge()</code> and <code>Onyx.set()</code>.</p>
 </dd>
-<dt><a href="#mergeCollection">mergeCollection(collectionKey, collection, mergeReplaceNullPatches)</a></dt>
-<dd><p>Merges a collection based on their keys</p>
+<dt><a href="#mergeCollection">mergeCollection(collectionKey, collection)</a></dt>
+<dd><p>Merges a collection based on their keys.</p>
 </dd>
 <dt><a href="#clear">clear(keysToPreserve)</a></dt>
 <dd><p>Clear out all the data in the store</p>
@@ -157,8 +157,8 @@ Onyx.merge(ONYXKEYS.POLICY, {name: 'My Workspace'}); // -> {id: 1, name: 'My Wor
 ```
 <a name="mergeCollection"></a>
 
-## mergeCollection(collectionKey, collection, mergeReplaceNullPatches)
-Merges a collection based on their keys
+## mergeCollection(collectionKey, collection)
+Merges a collection based on their keys.
 
 **Kind**: global function  
 
@@ -166,7 +166,6 @@ Merges a collection based on their keys
 | --- | --- |
 | collectionKey | e.g. `ONYXKEYS.COLLECTION.REPORT` |
 | collection | Object collection keyed by individual collection member keys and values |
-| mergeReplaceNullPatches | Record where the key is a collection member key and the value is a list of tuples that we'll use to replace the nested objects of that collection member record with something else. |
 
 **Example**  
 ```js
